requirements.txt
python-telegram-bot
openweathermap-api-client